workflow-resources/git-promotion/requirements.txt
GitPython==3.1.29
PyYAML==5.3.1
pytest